+++ app/assets/javascripts/view.js
@@ -63,11 +63,6 @@ var View = {
 
     $(document.body).click(this.dropdowns.removeFocus);
 
-    /* facebox */
-    $.facebox.settings.closeImage = app.baseImageUrl()+'facebox/closelabel.png';
-    $.facebox.settings.loadingImage = app.baseImageUrl()+'facebox/loading.gif';
-    $.facebox.settings.opacity = 0.75;
-
     $('a[rel*=facebox]').facebox();
     $(document).bind('reveal.facebox', function() {
       Diaspora.page.directionDetector.updateBinds();