+++ config/initializers/devise.rb
@@ -159,5 +159,8 @@ Devise.setup do |config|
   #   end
   #   manager.default_strategies(:scope => :user).unshift :twitter_oauth
   # end
+  
+  # Sign out via a DELETE request
+  config.sign_out_via = :delete
 end