+++ Changelog.md
@@ -3,6 +3,7 @@
 ## Refactor
 * Background actual mailing when sending invitations [#4069](https://github.com/diaspora/diaspora/issues/4069)
 * Set the current user on the client side through gon [#4028](https://github.com/diaspora/diaspora/issues/4028)
+* Update sign out route to a DELETE request [#4068](https://github.com/diaspora/diaspora/issues/4068)
 
 ## Bug fixes
 * Don't focus comment form on 'show n more comments' [#4265](https://github.com/diaspora/diaspora/issues/4265)
@@ -11,6 +12,8 @@
 * Disable invite button after sending invite [#4173](https://github.com/diaspora/diaspora/issues/4173)
 * Fix pagination for people list on the tag stream page [#4245](https://github.com/diaspora/diaspora/pull/4245)
 * Fix missing timeago tooltip in conversations [#4257](https://github.com/diaspora/diaspora/issues/4257)
+* Fix link to background image [#4289](https://github.com/diaspora/diaspora/pull/4289)
+* Fix Facebox icons 404s when called from Backbone
 
 ## Features
 * Admin: add option to find users under 13 (COPPA) [#4252](https://github.com/diaspora/diaspora/pull/4252)